add pipeline firstime
